Change all parts that use double for money: change them to use BigDecimal instead, logic stays the same

+++ src/main/java/com/auction/models/Admin.java
@@ -1,6 +1,7 @@
 package com.auction.models;
 
 import com.auction.server.factory.UserRole;
+import java.math.BigDecimal;
 
 public class Admin extends User {
     private static final long serialVersionUID = 1L;
@@ -11,7 +12,7 @@ public class Admin extends User {
     }
 
     // Create an Admin instance from an existed Admin in DB
-    public Admin(String id, String username, String password, double balance) {
+    public Admin(String id, String username, String password, BigDecimal balance) {
         super(id, UserRole.ADMIN, username, password, balance);
     }
 }
