+++ src/main/java/com/auction/server/factory/ItemFactory.java
@@ -1,12 +1,13 @@
 package com.auction.server.factory;
 
 import com.auction.models.*;
+import java.math.BigDecimal;
 
 /**
  * This class is used to create an Item object based on the specified ItemType.
  */
 public class ItemFactory {
-    public static Item createNewItem(ItemType type, String name, String description, double startingPrice,
+    public static Item createNewItem(ItemType type, String name, String description, BigDecimal startingPrice,
                                   String BrandOrArtist, Seller owner) throws IllegalArgumentException {
         switch (type) {
             case ELECTRONICS -> {
@@ -22,7 +23,7 @@ public class ItemFactory {
         }
     }
 
-    public static Item createItemFromDB(String id, ItemType type, String name, String description, double startingPrice,
+    public static Item createItemFromDB(String id, ItemType type, String name, String description, BigDecimal startingPrice,
                                         String BrandOrArtist, Seller owner) throws IllegalArgumentException {
         switch (type) {
             case ELECTRONICS -> {
